rename component using PascalCase
